Final 4ch to luppy replace, also URL fixes
The wordfilter was smart with regex and whatnot, but it could have left
bbehind some stuff that shouldn't be replaced. The new URLs are the
image and thumbnail URLs, also the thread URLs are lupchan specific
because nanasi has not written any nginx rewite rules to remove the .php
and .html file extensions.
Oh also I fixed some syntax in the post function.